app/pyproject.toml
[tool.poetry]
name = "emotive"
version = "0.1.0"
description = "Emotive tech test"
authors = ["Guilherme Tonello <guilherme.tonello21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.4"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
celery = "^5.4.0"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
redis = "^5.0.3"
pillow = "^10.3.0"

[tool.poetry.dev-dependencies]
black = "^24.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
